+++ src/Journalist.WindowsAzure.Storage.Abstractions/Tables/ICloudTableEntitySegmentedRangeQuery.cs
@@ -5,8 +5,12 @@ namespace Journalist.WindowsAzure.Storage.Tables
 {
     public interface ICloudTableEntitySegmentedRangeQuery
     {
+        Task<List<Dictionary<string, object>>> ExecuteAsync();
+
+        Task<List<Dictionary<string, object>>> ExecuteAsync(byte[] continuationToken);
+
         bool HasMore { get; }
 
-        Task<List<Dictionary<string, object>>> ExecuteAsync();
+        byte[] ContinuationToken { get; }
     }
 }